adapool integration
